requirements.txt
efinance>=0.5.5
akshare>=1.12.0
baostock>=0.8.0
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
exchange-calendars>=4.5
python-dotenv>=1.0
tenacity>=8.2
sqlalchemy>=2.0
rich>=13.0
click>=8.1
schedule>=1.2
openai>=1.0
pyyaml>=6.0
requests>=2.28
backtrader>=1.9.78.123
beautifulsoup4>=4.12
lxml>=4.9
flask>=2.0
fpdf2>=2.7
